refactor(gabinet): extract employees.$employeeId route components (closes #4452)
Move EmployeeNavGroups, AppointmentsTabContent, PatientsTabContent,
NotesTabContent, EditEmployeeDrawer, UpcomingAgenda, DetailedDataTab,
AssignedItemsTab, EmployeePermissionsTab, AccountTabContent, and
ChangePasswordDialog into src/components/gabinet/employees/.
Route file is now a thin shell.
